Extract four command renderers from cmd/render.go
Structural refactor item #9 from the consolidated review: cmd/render.go
was 2,068 LOC / 87 functions, a single switch-on-type god module that
every analysis command had to touch. This PR extracts the four largest
per-command renderers into their own files, a meaningful first step
toward the review's "one file per command" target.

New files, each containing a single command's renderer plus its
command-local helpers:

- cmd/render_status.go         (status command — renderStatusResult,
                                compact details, governance hotspots,
                                spec families, runtime-provider
                                summary, runtime-check glyph)
- cmd/render_docdrift.go       (check-doc-drift — renderDocDriftResult,
                                drift badge/summary/padding, drift item
                                and assessment helpers, remediation
                                lookup, review-side adapters)
- cmd/render_compliance.go     (check-compliance — renderComplianceResult,
                                finding group, top suggestions,
                                compliance badge, limiting-factor
                                humanizer)
- cmd/render_impact.go         (analyze-impact — renderAnalyzeImpactResult,
                                impact-summary target display,
                                review-side impact lines, top-N
                                slicers)

The dispatcher (renderCommandResult) and the ~13 smaller renderers stay
in cmd/render.go for now; follow-up PRs can finish the split.

No behavior change: every extracted function is moved verbatim, imports
are minimized per file, and all existing tests pass without
modification. Verified with make fmt, make vet, and go test ./... .

cmd/render.go: 2068 -> 1322 LOC (-36%).
