+++ cmd/render.go
@@ -323,285 +323,6 @@ func renderIndexSourceSummaries(w io.Writer, sources []source.LoadSourceSummary)
 	}
 }
 
-func renderStatusResult(w io.Writer, result *statusResult) {
-	p := presentationForWriter(w)
-	fmt.Fprintln(w, p.headerLine("status", ""))
-	fmt.Fprintln(w)
-
-	freshnessState := "unknown"
-	if result.Freshness != nil && result.Freshness.State != "" {
-		freshnessState = result.Freshness.State
-	}
-	fmt.Fprintf(
-		w,
-		"  %s specs  %s docs  %s chunks  %s  %s\n",
-		p.bold(fmt.Sprintf("%d", result.SpecCount)),
-		p.bold(fmt.Sprintf("%d", result.DocCount)),
-		p.bold(fmt.Sprintf("%d", result.ChunkCount)),
-		renderFreshnessLabel(p, freshnessState),
-		p.dim(statusEmbedderSummary(result.EmbedderProvider)),
-	)
-	fmt.Fprintln(w)
-	if result.Compact {
-		renderCompactStatusDetails(w, result)
-		return
-	}
-
-	if result.WorkspaceRoot != "" {
-		fmt.Fprintf(w, "  %s %s\n", p.dim("workspace:"), result.WorkspaceRoot)
-	}
-	fmt.Fprintf(w, "  %s %s\n", p.dim("config:"), result.ConfigPath)
-	if result.ConfigResolution != nil {
-		if result.ConfigResolution.Reason != "" {
-			fmt.Fprintf(w, "  %s %s\n", p.dim("config resolution:"), result.ConfigResolution.Reason)
-		}
-		if len(result.ConfigResolution.Candidates) > 0 {
-			fmt.Fprintf(w, "  %s\n", p.white("CONFIG CANDIDATES"))
-			for _, candidate := range result.ConfigResolution.Candidates {
-				fmt.Fprintf(w, "  %s %d. %s | %s", p.treeItem(candidate.Precedence == len(result.ConfigResolution.Candidates)), candidate.Precedence, configSourceLabel(candidate.Source), candidate.Status)
-				if candidate.Path != "" {
-					fmt.Fprintf(w, " | %s", candidate.Path)
-				}
-				fmt.Fprintln(w)
-				if candidate.Detail != "" {
-					fmt.Fprintf(w, "     %s\n", p.dim(candidate.Detail))
-				}
-			}
-		}
-	}
-	fmt.Fprintf(w, "  %s %s\n", p.dim("index path:"), result.IndexPath)
-	if result.IndexExists {
-		fmt.Fprintf(w, "  %s %s\n", p.dim("index:"), p.green("present"))
-	} else {
-		fmt.Fprintf(w, "  %s %s\n", p.dim("index:"), p.red("missing"))
-	}
-	if len(result.Repos) > 0 {
-		fmt.Fprintf(w, "  %s\n", p.white("REPO COVERAGE"))
-		for i, repo := range result.Repos {
-			fmt.Fprintf(w, "  %s %s\n", p.treeItem(i == len(result.Repos)-1), renderRepoCoverageLine(repo))
-		}
-	}
-	if result.Freshness != nil {
-		fmt.Fprintf(w, "  %s %s\n", p.dim("index freshness:"), renderFreshnessLabel(p, result.Freshness.State))
-		for _, issue := range result.Freshness.Issues {
-			fmt.Fprintf(w, "  %s %s\n", p.cross(), issue.Message)
-		}
-		if result.Freshness.Action != "" {
-			fmt.Fprintf(w, "  %s %s\n", p.arrow(), result.Freshness.Action)
-		}
-	}
-	if result.RelationGraph != nil {
-		fmt.Fprintf(w, "  %s %s\n", p.dim("relation graph:"), result.RelationGraph.State)
-		for _, finding := range result.RelationGraph.Findings {
-			fmt.Fprintf(w, "  %s %s\n", p.cross(), finding.Message)
-		}
-	}
-	renderStatusGovernanceHotspots(w, result.GovernanceHotspots)
-	if result.ArtifactLocations != nil {
-		fmt.Fprintf(w, "  %s %s\n", p.dim("artifact index dir:"), result.ArtifactLocations.IndexDir)
-		fmt.Fprintf(w, "  %s %s\n", p.dim("artifact discover --write default:"), result.ArtifactLocations.DiscoverConfigPath)
-		fmt.Fprintf(w, "  %s %s\n", p.dim("artifact canonicalize default:"), result.ArtifactLocations.CanonicalizeBundleRoot)
-		if len(result.ArtifactLocations.IgnorePatterns) > 0 {
-			fmt.Fprintf(w, "  %s %s\n", p.dim("artifact ignore patterns:"), strings.Join(result.ArtifactLocations.IgnorePatterns, ", "))
-		}
-		for _, hint := range result.ArtifactLocations.RelocationHints {
-			fmt.Fprintf(w, "  %s %s\n", p.arrow(), hint)
-		}
-	}
-	if result.RuntimeConfig != nil {
-		fmt.Fprintf(w, "  %s\n", p.white("RUNTIME CONFIG"))
-		for i, item := range []struct {
-			Name     string
-			Provider statusRuntimeProvider
-		}{
-			{Name: "runtime.embedder", Provider: result.RuntimeConfig.Embedder},
-			{Name: "runtime.analysis", Provider: result.RuntimeConfig.Analysis},
-		} {
-			fmt.Fprintf(w, "  %s %s\n", p.treeItem(i == 1), renderRuntimeProviderSummary(item.Name, item.Provider))
-		}
-	}
-	if result.Runtime != nil {
-		fmt.Fprintf(w, "  %s %s\n", p.dim("runtime probe:"), result.Runtime.Scope)
-		for _, check := range result.Runtime.Checks {
-			fmt.Fprintf(w, "  %s %s %s | %s | provider: %s", p.dim("runtime:"), runtimeCheckGlyph(p, check.Status), check.Name, check.Status, check.Provider)
-			if check.Profile != "" {
-				fmt.Fprintf(w, " | profile: %s", check.Profile)
-			}
-			if check.Model != "" {
-				fmt.Fprintf(w, " | model: %s", check.Model)
-			}
-			if check.Endpoint != "" {
-				fmt.Fprintf(w, " | endpoint: %s", check.Endpoint)
-			}
-			if check.Timeout > 0 {
-				fmt.Fprintf(w, " | timeout_ms: %d", check.Timeout)
-			}
-			if check.Retries > 0 {
-				fmt.Fprintf(w, " | max_retries: %d", check.Retries)
-			}
-			fmt.Fprintln(w)
-			if check.Message != "" {
-				fmt.Fprintf(w, "     %s %s\n", p.dim("runtime note:"), check.Message)
-			}
-		}
-	}
-	renderSpecFamilies(w, result.Families)
-	for _, guidance := range result.Guidance {
-		fmt.Fprintf(w, "  %s %s\n", p.arrow(), guidance)
-	}
-}
-
-func renderCompactStatusDetails(w io.Writer, result *statusResult) {
-	p := presentationForWriter(w)
-	if result.Runtime != nil {
-		summaries := make([]string, 0, len(result.Runtime.Checks))
-		for _, check := range result.Runtime.Checks {
-			summaries = append(summaries, fmt.Sprintf("%s %s", check.Name, check.Status))
-		}
-		if len(summaries) > 0 {
-			fmt.Fprintf(w, "  %s %s | %s\n", p.dim("runtime probe:"), result.Runtime.Scope, strings.Join(summaries, ", "))
-		}
-	}
-	if result.Freshness != nil {
-		for _, issue := range result.Freshness.Issues {
-			fmt.Fprintf(w, "  %s %s\n", p.cross(), issue.Message)
-		}
-		if result.Freshness.Action != "" {
-			fmt.Fprintf(w, "  %s %s\n", p.arrow(), result.Freshness.Action)
-		}
-	}
-	if result.RelationGraph != nil {
-		for _, finding := range result.RelationGraph.Findings {
-			fmt.Fprintf(w, "  %s %s\n", p.cross(), finding.Message)
-		}
-	}
-	for _, guidance := range result.Guidance {
-		fmt.Fprintf(w, "  %s %s\n", p.arrow(), guidance)
-	}
-}
-
-func renderStatusGovernanceHotspots(w io.Writer, hotspots *index.GovernanceHotspots) {
-	lines := statusGovernanceHotspotLines(hotspots)
-	if len(lines) == 0 {
-		return
-	}
-
-	p := presentationForWriter(w)
-	fmt.Fprintf(w, "  %s\n", p.white("GOVERNANCE HOTSPOTS"))
-	for i, line := range lines {
-		fmt.Fprintf(w, "  %s %s\n", p.treeItem(i == len(lines)-1), line)
-	}
-}
-
-func statusGovernanceHotspotLines(hotspots *index.GovernanceHotspots) []string {
-	if hotspots == nil {
-		return nil
-	}
-
-	lines := make([]string, 0, len(hotspots.HighFanOutSpecs)+len(hotspots.WeakLinkArtifacts)+len(hotspots.MultiGovernedArtifacts))
-	for _, hotspot := range hotspots.HighFanOutSpecs {
-		line := fmt.Sprintf("fan-out spec %s", hotspot.Ref)
-		if hotspot.Title != "" {
-			line += " · " + hotspot.Title
-		}
-		line += fmt.Sprintf(" | %d applies_to edges", hotspot.AppliesToCount)
-		if weakEdges := hotspot.InferredEdgeCount + hotspot.AmbiguousEdgeCount; weakEdges > 0 {
-			line += fmt.Sprintf(" | %d weak edges", weakEdges)
-		}
-		lines = append(lines, line)
-	}
-	for _, hotspot := range hotspots.WeakLinkArtifacts {
-		line := fmt.Sprintf("weak-link artifact %s | %d governing specs", hotspot.Ref, hotspot.GoverningSpecCount)
-		if hotspot.SourceRef != "" && hotspot.SourceRef != hotspot.Ref {
-			line += " | " + displaySourcePath(hotspot.SourceRef)
-		}
-		if hotspot.InferredEdgeCount > 0 {
-			line += fmt.Sprintf(" | %d inferred", hotspot.InferredEdgeCount)
-		}
-		if hotspot.AmbiguousEdgeCount > 0 {
-			line += fmt.Sprintf(" | %d ambiguous", hotspot.AmbiguousEdgeCount)
-		}
-		if specs := formatGovernanceHotspotSpecRefs(hotspot.GoverningSpecs, 3); specs != "" {
-			line += " | " + specs
-		}
-		lines = append(lines, line)
-	}
-	for _, hotspot := range hotspots.MultiGovernedArtifacts {
-		line := fmt.Sprintf("multi-governed artifact %s | %d governing specs", hotspot.Ref, hotspot.GoverningSpecCount)
-		if hotspot.SourceRef != "" && hotspot.SourceRef != hotspot.Ref {
-			line += " | " + displaySourcePath(hotspot.SourceRef)
-		}
-		if weakEdges := hotspot.InferredEdgeCount + hotspot.AmbiguousEdgeCount; weakEdges > 0 {
-			line += fmt.Sprintf(" | %d weak edges", weakEdges)
-		}
-		if specs := formatGovernanceHotspotSpecRefs(hotspot.GoverningSpecs, 3); specs != "" {
-			line += " | " + specs
-		}
-		lines = append(lines, line)
-	}
-	return lines
-}
-
-func formatGovernanceHotspotSpecRefs(refs []string, limit int) string {
-	if len(refs) == 0 {
-		return ""
-	}
-	if limit <= 0 || len(refs) <= limit {
-		return strings.Join(refs, ", ")
-	}
-	return fmt.Sprintf("%s +%d more", strings.Join(refs[:limit], ", "), len(refs)-limit)
-}
-
-func renderSpecFamilies(w io.Writer, families *index.FamilyResult) {
-	if families == nil || len(families.Families) == 0 {
-		return
-	}
-	p := presentationForWriter(w)
-	fmt.Fprintf(w, "  %s\n", p.white("SPEC FAMILIES"))
-	for i, family := range families.Families {
-		memberList := strings.Join(family.Members, ", ")
-		fmt.Fprintf(w, "  %s family %d (%d member(s), cohesion: %.2f): %s\n",
-			p.treeItem(i == len(families.Families)-1 && len(families.Ungoverned) == 0),
-			family.ID, family.Size, family.Cohesion, memberList)
-	}
-	if len(families.Ungoverned) > 0 {
-		fmt.Fprintf(w, "  %s ungoverned files: %d\n", p.dim("coverage gap:"), len(families.Ungoverned))
-		limit := 5
-		if len(families.Ungoverned) < limit {
-			limit = len(families.Ungoverned)
-		}
-		for _, path := range families.Ungoverned[:limit] {
-			fmt.Fprintf(w, "    %s %s\n", p.cross(), path)
-		}
-		if len(families.Ungoverned) > 5 {
-			fmt.Fprintf(w, "    %s ... and %d more\n", p.dim(""), len(families.Ungoverned)-5)
-		}
-	}
-}
-
-func renderRuntimeProviderSummary(name string, provider statusRuntimeProvider) string {
-	parts := []string{name}
-	if provider.Profile != "" {
-		parts = append(parts, "profile: "+provider.Profile)
-	}
-	if provider.Provider != "" {
-		parts = append(parts, "provider: "+provider.Provider)
-	}
-	if provider.Model != "" {
-		parts = append(parts, "model: "+provider.Model)
-	}
-	if provider.Endpoint != "" {
-		parts = append(parts, "endpoint: "+provider.Endpoint)
-	}
-	if provider.TimeoutMS > 0 {
-		parts = append(parts, fmt.Sprintf("timeout_ms: %d", provider.TimeoutMS))
-	}
-	if provider.MaxRetries > 0 {
-		parts = append(parts, fmt.Sprintf("max_retries: %d", provider.MaxRetries))
-	}
-	return strings.Join(parts, " | ")
-}
-
 func renderVersionResult(w io.Writer, result *versionResult) {
 	fmt.Fprintf(w, "version: %s\n", result.Version)
 	fmt.Fprintf(w, "go version: %s\n", result.GoVersion)
@@ -934,194 +655,6 @@ func renderCompareResult(w io.Writer, result *analysis.CompareResult) {
 	}
 }
 
-func renderAnalyzeImpactResult(w io.Writer, result *analysis.AnalyzeImpactResult) {
-	fmt.Fprintf(w, "spec: %s | change_type: %s\n", result.SpecRef, result.ChangeType)
-	if len(result.RankedSummary) > 0 {
-		fmt.Fprintf(w, "ranked summary: %d\n", len(result.RankedSummary))
-		for _, item := range result.RankedSummary {
-			fmt.Fprintf(w, "%d. %s %s", item.Rank, item.Kind, item.Ref)
-			if item.Repo != "" {
-				fmt.Fprintf(w, " | repo: %s", item.Repo)
-			}
-			if item.SourceRef != "" {
-				fmt.Fprintf(w, " | source: %s", displaySourcePath(item.SourceRef))
-			}
-			if item.Score > 0 {
-				fmt.Fprintf(w, " | %.3f", item.Score)
-			}
-			if item.Title != "" {
-				fmt.Fprintf(w, " | %s", item.Title)
-			}
-			if item.Why != "" {
-				fmt.Fprintf(w, " | why: %s", item.Why)
-			}
-			if item.ReviewFirst != "" {
-				fmt.Fprintf(w, " | review first: %s", displayImpactSummaryTarget(item.ReviewFirst))
-			}
-			fmt.Fprintln(w)
-		}
-	}
-	if result.SummaryOnly {
-		return
-	}
-	fmt.Fprintf(w, "affected specs: %d\n", len(result.AffectedSpecs))
-	for _, item := range result.AffectedSpecs {
-		fmt.Fprintf(w, "- %s", item.Ref)
-		if item.Repo != "" {
-			fmt.Fprintf(w, " | repo: %s", item.Repo)
-		}
-		fmt.Fprintf(w, " | %s", item.Relationship)
-		if item.Historical {
-			fmt.Fprint(w, " | historical")
-		}
-		if item.Title != "" {
-			fmt.Fprintf(w, " | %s", item.Title)
-		}
-		fmt.Fprintln(w)
-	}
-	fmt.Fprintf(w, "affected refs: %d\n", len(result.AffectedRefs))
-	for _, item := range result.AffectedRefs {
-		fmt.Fprintf(w, "- %s | %s\n", item.Ref, item.Kind)
-	}
-	fmt.Fprintf(w, "affected docs: %d\n", len(result.AffectedDocs))
-	for _, item := range result.AffectedDocs {
-		fmt.Fprintf(w, "- %s", item.Ref)
-		if item.Repo != "" {
-			fmt.Fprintf(w, " | repo: %s", item.Repo)
-		}
-		if item.SourceRef != "" {
-			fmt.Fprintf(w, " | source: %s", displaySourcePath(item.SourceRef))
-		}
-		fmt.Fprintf(w, " | %.3f", item.Score)
-		if item.Classification != "" {
-			fmt.Fprintf(w, " | %s", item.Classification)
-		}
-		if item.Title != "" {
-			fmt.Fprintf(w, " | %s", item.Title)
-		}
-		fmt.Fprintln(w)
-		if len(item.Reasons) > 0 {
-			fmt.Fprintf(w, "  reason: %s\n", item.Reasons[0])
-		}
-		if item.Evidence != nil {
-			fmt.Fprintf(
-				w,
-				"  evidence: %s / %s -> %s / %s\n",
-				displaySourcePath(item.Evidence.SpecSourceRef),
-				displayDefault(item.Evidence.SpecSection, "(body)"),
-				displaySourcePath(item.Evidence.DocSourceRef),
-				displayDefault(item.Evidence.DocSection, "(body)"),
-			)
-			if strings.TrimSpace(item.Evidence.LinkReason) != "" {
-				fmt.Fprintf(w, "  why: %s\n", item.Evidence.LinkReason)
-			}
-		}
-		if len(item.SuggestedTargets) > 0 {
-			target := item.SuggestedTargets[0]
-			fmt.Fprintf(w, "  target: %s", displaySourcePath(target.SourceRef))
-			if strings.TrimSpace(target.Section) != "" {
-				fmt.Fprintf(w, " / %s", target.Section)
-			}
-			fmt.Fprintln(w)
-		}
-	}
-}
-
-func displayImpactSummaryTarget(value string) string {
-	head, tail, ok := strings.Cut(value, " / ")
-	if !ok {
-		return displaySourcePath(value)
-	}
-	return displaySourcePath(head) + " / " + tail
-}
-
-func renderComplianceResult(w io.Writer, result *analysis.ComplianceResult) {
-	p := presentationForWriter(w)
-	fmt.Fprintln(w, p.headerLine("check-compliance", ""))
-	fmt.Fprintln(w)
-	if len(result.Paths) > 0 {
-		fmt.Fprintf(w, "  %s %s\n", p.dim("paths:"), strings.Join(result.Paths, ", "))
-	}
-	fmt.Fprintf(w, "  %s %d\n", p.dim("relevant specs:"), len(result.RelevantSpecs))
-	fmt.Fprintln(w)
-	renderComplianceFindingGroup(w, "conflicts", result.Conflicts)
-	renderComplianceFindingGroup(w, "compliant", result.Compliant)
-	renderComplianceFindingGroup(w, "unspecified", result.Unspecified)
-	if suggestions := complianceTopSuggestions(result); len(suggestions) > 0 {
-		fmt.Fprintln(w)
-		fmt.Fprintf(w, "  %s %d\n", p.white("TOP SUGGESTIONS:"), len(suggestions))
-		for _, suggestion := range suggestions {
-			fmt.Fprintf(w, "     %s %s\n", p.arrow(), truncateRenderLine(suggestion, 160))
-		}
-	}
-}
-
-func renderComplianceFindingGroup(w io.Writer, label string, findings []analysis.ComplianceFinding) {
-	p := presentationForWriter(w)
-	if len(findings) == 0 {
-		fmt.Fprintf(w, "  %s none\n", p.dim(label+":"))
-		return
-	}
-
-	fmt.Fprintf(w, "  %s %d\n", p.white(strings.ToUpper(label)+":"), len(findings))
-	for _, item := range findings {
-		fmt.Fprintf(w, "  %s %s", complianceBadge(p, label), item.Path)
-		if item.SpecRef != "" {
-			fmt.Fprintf(w, " | %s", p.cyan(item.SpecRef))
-		}
-		if item.SectionHeading != "" {
-			fmt.Fprintf(w, " | %s", item.SectionHeading)
-		}
-		fmt.Fprintf(w, " | %s\n", item.Message)
-		if item.Traceability != "" {
-			fmt.Fprintf(w, "     %s %s\n", p.dim("traceability"), item.Traceability)
-		}
-		if item.LimitingFactor != "" {
-			fmt.Fprintf(w, "     %s %s\n", p.dim("limiting factor"), humanizeComplianceLimitingFactor(item.LimitingFactor))
-		}
-		if item.Suggestion != "" {
-			fmt.Fprintf(w, "     %s %s\n", p.arrow(), item.Suggestion)
-		}
-	}
-}
-
-func complianceTopSuggestions(result *analysis.ComplianceResult) []string {
-	if result == nil {
-		return nil
-	}
-	if len(result.TopSuggestions) > 0 {
-		return result.TopSuggestions
-	}
-
-	seen := make(map[string]struct{})
-	suggestions := make([]string, 0, 3)
-	appendSuggestions := func(findings []analysis.ComplianceFinding) {
-		for _, finding := range findings {
-			suggestion := strings.TrimSpace(finding.Suggestion)
-			if suggestion == "" {
-				continue
-			}
-			if _, ok := seen[suggestion]; ok {
-				continue
-			}
-			seen[suggestion] = struct{}{}
-			suggestions = append(suggestions, suggestion)
-			if len(suggestions) == 3 {
-				return
-			}
-		}
-	}
-
-	appendSuggestions(result.Conflicts)
-	if len(suggestions) < 3 {
-		appendSuggestions(result.Unspecified)
-	}
-	if len(suggestions) < 3 {
-		appendSuggestions(result.Compliant)
-	}
-	return suggestions
-}
-
 func truncateRenderLine(value string, limit int) string {
 	value = strings.TrimSpace(value)
 	if limit <= 0 || len(value) <= limit {
@@ -1199,106 +732,6 @@ func renderTerminologyFinding(w io.Writer, label string, finding analysis.Termin
 			if section.Evidence.Excerpt != "" {
 				fmt.Fprintf(w, "     expected: %s\n", section.Evidence.Excerpt)
 			}
-		}
-	}
-}
-
-func renderDocDriftResult(w io.Writer, result *analysis.DocDriftResult) {
-	p := presentationForWriter(w)
-	fmt.Fprintln(w, p.headerLine("check-doc-drift", ""))
-	fmt.Fprintln(w)
-
-	if len(result.DriftItems) == 0 && len(result.Assessments) == 0 && len(result.ChangedFiles) == 0 && len(result.ImplicatedSpecs) == 0 && len(result.ImplicatedDocs) == 0 {
-		fmt.Fprintf(w, "  %s no drift items\n", p.check())
-		return
-	}
-	if len(result.ChangedFiles) > 0 {
-		fmt.Fprintf(w, "  %s %d\n", p.white("CHANGED FILES"), len(result.ChangedFiles))
-		for i, file := range result.ChangedFiles {
-			fmt.Fprintf(w, "  %s %s", p.treeBranch(i == len(result.ChangedFiles)-1), file.Path)
-			if file.AddedLineCount > 0 || file.RemovedLineCount > 0 {
-				fmt.Fprintf(w, " %s", p.dim(fmt.Sprintf("(+%d/-%d)", file.AddedLineCount, file.RemovedLineCount)))
-			}
-			fmt.Fprintln(w)
-		}
-		fmt.Fprintln(w)
-	}
-	if len(result.ImplicatedSpecs) > 0 {
-		fmt.Fprintf(w, "  %s %d\n", p.white("IMPLICATED SPECS"), len(result.ImplicatedSpecs))
-		for i, item := range result.ImplicatedSpecs {
-			line := item.Ref
-			if item.Title != "" {
-				line += " · " + item.Title
-			}
-			if item.Repo != "" {
-				line += " · repo: " + item.Repo
-			}
-			if len(item.Reasons) > 0 {
-				line += " · " + item.Reasons[0]
-			}
-			fmt.Fprintf(w, "  %s %s\n", p.treeBranch(i == len(result.ImplicatedSpecs)-1), line)
-		}
-		fmt.Fprintln(w)
-	}
-	if len(result.ImplicatedDocs) > 0 {
-		fmt.Fprintf(w, "  %s %d\n", p.white("IMPLICATED DOCS"), len(result.ImplicatedDocs))
-		for i, item := range result.ImplicatedDocs {
-			label := repoPathLabel(item.Repo, preferredDocLabel(item.DocRef, item.SourceRef))
-			line := fmt.Sprintf("%s  %.3f", label, item.Score)
-			if len(item.Reasons) > 0 {
-				line += " · " + item.Reasons[0]
-			}
-			fmt.Fprintf(w, "  %s %s\n", p.treeBranch(i == len(result.ImplicatedDocs)-1), line)
-		}
-		fmt.Fprintln(w)
-	}
-
-	assessments := result.Assessments
-	if len(assessments) == 0 {
-		assessments = driftAssessmentsFromItems(result.DriftItems)
-	}
-	driftItems := driftItemsByDocRef(result.DriftItems)
-	remediation := remediationItemsByDocRef(result.Remediation)
-	for i, assessment := range assessments {
-		if i > 0 {
-			fmt.Fprintln(w)
-		}
-		docLabel := repoPathLabel(assessment.Repo, preferredDocLabel(assessment.DocRef, assessment.SourceRef))
-		fmt.Fprintf(w, "  %s", p.cyan(docLabel))
-		if padding := docDriftPadding(docLabel); padding > 0 {
-			fmt.Fprint(w, strings.Repeat(" ", padding))
-		} else {
-			fmt.Fprint(w, "  ")
-		}
-		fmt.Fprintln(w, driftAssessmentBadge(p, assessment.Status))
-		if item, ok := driftItems[assessment.DocRef]; ok && len(item.Findings) > 0 {
-			for _, finding := range item.Findings {
-				fmt.Fprintf(w, "\n    %s %s", p.cross(), p.bold(driftFindingSummary(finding)))
-				if finding.Expected != "" {
-					fmt.Fprintf(w, "  %s %s", p.yellow("expected"), finding.Expected)
-				}
-				if finding.Observed != "" {
-					fmt.Fprintf(w, "  %s %s", p.yellow("got"), finding.Observed)
-				}
-				fmt.Fprintln(w)
-			}
-		} else if assessment.Rationale != "" {
-			fmt.Fprintf(w, "\n    %s %s\n", p.arrow(), assessment.Rationale)
-		}
-		if suggestions := remediation[assessment.DocRef]; len(suggestions) > 0 {
-			pathArg := preferredDocLabel(assessment.DocRef, assessment.SourceRef)
-			if assessment.SourceRef != "" {
-				pathArg = displaySourcePath(assessment.SourceRef)
-			}
-			if isNonPrimaryRepoDoc(assessment.DocRef, assessment.Repo) {
-				fmt.Fprintf(w, "\n    %s  deterministic remediation is available, but `pituitary fix --path` only targets primary-workspace docs; inspect %s manually\n", p.info(), p.cyan(docLabel))
-				fmt.Fprintf(w, "    %s  run `pituitary review-spec --format html --path <spec>` for the full evidence report\n", p.info())
-			} else {
-				fmt.Fprintf(w, "\n    %s pituitary fix --path %s %s\n", p.green("fix:"), pathArg, p.dim(fmt.Sprintf("(%d edits)", len(suggestions))))
-				fmt.Fprintf(w, "    %s  run `pituitary review-spec --format html --path <spec>` for the full evidence report\n", p.info())
-			}
-		} else if assessment.Status == "drift" || assessment.Status == "possible_drift" {
-			fmt.Fprintf(w, "\n    %s  run `pituitary review-spec --format html --path <spec>` for the full evidence chain (no deterministic fix available)\n", p.info())
 		}
 	}
 }
@@ -1633,34 +1066,6 @@ func reviewMarkdownActions(result *analysis.ReviewResult) []string {
 	return actions
 }
 
-func reviewDocDriftAssessments(result *analysis.DocDriftResult) []analysis.DocDriftAssessment {
-	if result == nil {
-		return nil
-	}
-	if len(result.Assessments) > 0 {
-		items := make([]analysis.DocDriftAssessment, 0, len(result.Assessments))
-		for _, item := range result.Assessments {
-			if item.Status == "aligned" {
-				continue
-			}
-			items = append(items, item)
-		}
-		return items
-	}
-	return driftAssessmentsFromItems(result.DriftItems)
-}
-
-func reviewRemediationSuggestionCount(result *analysis.DocRemediationResult) int {
-	if result == nil {
-		return 0
-	}
-	count := 0
-	for _, item := range result.Items {
-		count += len(item.Suggestions)
-	}
-	return count
-}
-
 func renderFreshnessLabel(p renderPresentation, state string) string {
 	switch state {
 	case "fresh":
@@ -1677,28 +1082,6 @@ func statusEmbedderSummary(provider string) string {
 		return "embedder unknown"
 	}
 	return provider + " embedder"
-}
-
-func runtimeCheckGlyph(p renderPresentation, status string) string {
-	switch status {
-	case "ready":
-		return p.check()
-	case "disabled":
-		return p.info()
-	default:
-		return p.cross()
-	}
-}
-
-func complianceBadge(p renderPresentation, label string) string {
-	switch label {
-	case "conflicts":
-		return p.conflictBadge()
-	case "compliant":
-		return p.compliantBadge()
-	default:
-		return p.unspecifiedBadge()
-	}
 }
 
 func preferredDocLabel(docRef, sourceRef string) string {
@@ -1745,36 +1128,6 @@ func isNonPrimaryRepoDoc(docRef, repo string) bool {
 	return strings.HasPrefix(strings.TrimSpace(docRef), "doc://"+repo+"/")
 }
 
-func docDriftPadding(label string) int {
-	width := 56 - len([]rune(label))
-	if width < 2 {
-		return 2
-	}
-	return width
-}
-
-func driftAssessmentBadge(p renderPresentation, status string) string {
-	switch status {
-	case "aligned":
-		return p.okBadge()
-	case "possible_drift":
-		return p.reviewBadge()
-	default:
-		return p.driftBadge()
-	}
-}
-
-func driftFindingSummary(finding analysis.DriftFinding) string {
-	summary := strings.TrimSpace(finding.Message)
-	if strings.TrimSpace(finding.Code) != "" {
-		summary = humanizeSymbol(finding.Code)
-	}
-	if finding.Classification == "role_mismatch" {
-		return "role mismatch · " + summary
-	}
-	return summary
-}
-
 func overlapBlock(p renderPresentation, item analysis.OverlapItem) string {
 	if item.Guidance == "merge_candidate" || item.Score >= 0.85 {
 		return p.blockHigh()
@@ -1798,40 +1151,6 @@ func pluralSuffix(count int) string {
 		return ""
 	}
 	return "s"
-}
-
-func reviewImpactLines(result *analysis.AnalyzeImpactResult) []string {
-	if result == nil {
-		return nil
-	}
-	lines := make([]string, 0, 6)
-	for _, item := range topImpactedSpecs(result.AffectedSpecs, 2) {
-		line := fmt.Sprintf("%s  %s · %s", item.Ref, item.Title, item.Relationship)
-		if item.Repo != "" {
-			line += " · repo: " + item.Repo
-		}
-		if item.Historical {
-			line += " · historical"
-		}
-		lines = append(lines, line)
-	}
-	for _, item := range topImpactedDocs(result.AffectedDocs, 2) {
-		line := fmt.Sprintf("%s  %.3f", item.Ref, item.Score)
-		if item.Classification != "" {
-			line += " · " + item.Classification
-		}
-		if item.Repo != "" {
-			line += " · repo: " + item.Repo
-		}
-		if item.SourceRef != "" {
-			line += " · " + displaySourcePath(item.SourceRef)
-		}
-		if len(item.SuggestedTargets) > 0 && item.SuggestedTargets[0].Section != "" {
-			line += " · target: " + item.SuggestedTargets[0].Section
-		}
-		lines = append(lines, line)
-	}
-	return lines
 }
 
 func humanizeSymbol(value string) string {
@@ -1917,71 +1236,6 @@ func humanizeOverlapGuidance(guidance string) string {
 	default:
 		return guidance
 	}
-}
-
-func humanizeComplianceLimitingFactor(factor string) string {
-	switch factor {
-	case "spec_metadata_gap":
-		return "accepted spec metadata is missing explicit applies_to coverage"
-	case "code_evidence_gap":
-		return "the file or diff does not expose enough literal code evidence to confirm compliance"
-	default:
-		return factor
-	}
-}
-
-func remediationItemsByDocRef(result *analysis.DocRemediationResult) map[string][]analysis.DocRemediationSuggestion {
-	if result == nil {
-		return map[string][]analysis.DocRemediationSuggestion{}
-	}
-	items := make(map[string][]analysis.DocRemediationSuggestion, len(result.Items))
-	for _, item := range result.Items {
-		items[item.DocRef] = item.Suggestions
-	}
-	return items
-}
-
-func driftItemsByDocRef(items []analysis.DriftItem) map[string]analysis.DriftItem {
-	result := make(map[string]analysis.DriftItem, len(items))
-	for _, item := range items {
-		result[item.DocRef] = item
-	}
-	return result
-}
-
-func driftAssessmentsFromItems(items []analysis.DriftItem) []analysis.DocDriftAssessment {
-	result := make([]analysis.DocDriftAssessment, 0, len(items))
-	for _, item := range items {
-		result = append(result, analysis.DocDriftAssessment{
-			DocRef:    item.DocRef,
-			Title:     item.Title,
-			Repo:      item.Repo,
-			SourceRef: item.SourceRef,
-			Status:    "drift",
-			SpecRefs:  append([]string(nil), item.SpecRefs...),
-		})
-	}
-	return result
-}
-
-func topImpactedSpecs(items []analysis.ImpactedSpec, limit int) []analysis.ImpactedSpec {
-	if len(items) == 0 || limit <= 0 {
-		return nil
-	}
-	if len(items) <= limit {
-		return append([]analysis.ImpactedSpec(nil), items...)
-	}
-	return append([]analysis.ImpactedSpec(nil), items[:limit]...)
-}
-
-func topImpactedDocs(items []analysis.ImpactedDoc, limit int) []analysis.ImpactedDoc {
-	if len(items) == 0 || limit <= 0 {
-		return nil
-	}
-	if len(items) <= limit {
-		return append([]analysis.ImpactedDoc(nil), items...)
-	}
-	return append([]analysis.ImpactedDoc(nil), items[:limit]...)
 }
 
 func minInt(left, right int) int {
